# investment-system/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 投资管理系统统一接口契约
  version: 1.1.0
  description: |
    LTC 投资管理系统正式接口契约。认证使用 HttpOnly Cookie，会话写操作同时校验 X-CSRF-Token。
    所有投资业务单据统一按“经办人、经办部门负责人、经办部门分管领导”顺序流转。
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: 本地及部署环境 API 基地址
security:
  - cookieAuth: []
...
paths:
...
  /admin/units:
    get:
      operationId: listAdminUnits
      summary: 分页查询后台单位
      x-permissions: [admin.unit.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses: *okResponses
    post:
      operationId: createAdminUnit
      summary: 创建单位
      x-permissions: [admin.unit.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitInput'
      responses: *createdResponses
  /admin/units/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      operationId: updateAdminUnit
      summary: 修改单位
      x-permissions: [admin.unit.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitUpdateInput'
      responses: *okResponses
    delete:
      operationId: deleteAdminUnit
      summary: 删除未被引用的单位
      x-permissions: [admin.unit.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses: *okResponses
  /admin/departments:
    get:
      operationId: listAdminDepartments
      summary: 分页查询后台部门
      x-permissions: [admin.department.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses: *okResponses
    post:
      operationId: createAdminDepartment
      summary: 创建部门并配置审批负责人
      x-permissions: [admin.department.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentInput'
      responses: *createdResponses
  /admin/departments/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      operationId: updateAdminDepartment
      summary: 修改部门及审批负责人配置
      x-permissions: [admin.department.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentUpdateInput'
      responses: *okResponses
    delete:
      operationId: deleteAdminDepartment
      summary: 删除未被引用的部门
      x-permissions: [admin.department.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses: *okResponses
  /admin/positions:
    get:
      operationId: listAdminPositions
      summary: 分页查询后台岗位
      x-permissions: [admin.position.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses: *okResponses
    post:
      operationId: createAdminPosition
      summary: 创建岗位
      x-permissions: [admin.position.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionInput'
      responses: *createdResponses
  /admin/positions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      operationId: updateAdminPosition
      summary: 修改岗位
      x-permissions: [admin.position.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionUpdateInput'
      responses: *okResponses
    delete:
      operationId: deleteAdminPosition
      summary: 删除未被引用的岗位
      x-permissions: [admin.position.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses: *okResponses
  /admin/permissions:
    get:
      operationId: listPermissions
      summary: 查询系统固定权限点
      x-permissions: [admin.role.read]
      responses: *okResponses
  /admin/roles:
    get:
      operationId: listAdminRoles
      summary: 分页查询角色及授权
      x-permissions: [admin.role.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses: *okResponses
    post:
      operationId: createAdminRole
      summary: 创建角色并分配权限
      x-permissions: [admin.role.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses: *createdResponses
  /admin/roles/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      operationId: updateAdminRole
      summary: 修改角色及权限
      x-permissions: [admin.role.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateInput'
      responses: *okResponses
    delete:
      operationId: deleteAdminRole
      summary: 删除非系统角色
      x-permissions: [admin.role.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses: *okResponses
  /admin/users:
    get:
      operationId: listAdminUsers
      summary: 分页查询人员账号
      x-permissions: [admin.user.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses: *okResponses
    post:
      operationId: createAdminUser
      summary: 创建人员账号并分配角色
      x-permissions: [admin.user.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses: *createdResponses
  /admin/users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      operationId: updateAdminUser
      summary: 修改人员、组织归属、状态和角色
      x-permissions: [admin.user.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateInput'
      responses: *okResponses
    delete:
      operationId: deleteAdminUser
      summary: 删除未被业务引用的人员账号
      x-permissions: [admin.user.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses: *okResponses
  /admin/users/{id}/reset-password:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      operationId: resetAdminUserPassword
      summary: 重置人员密码并撤销其会话
      x-permissions: [admin.user.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordInput'
      responses: *okResponses
  /directory/units:
    get:
      operationId: listDirectoryUnits
      summary: 查询业务表单可选择的启用单位
      x-permissions: [investment.intention.create, investment.intention.update, investment.group_decision.create, investment.group_decision.update, admin.unit.read]
      responses: *okResponses
  /directory/users:
    get:
      operationId: listDirectoryUsers
      summary: 查询业务表单可选择的启用人员
      x-permissions: [investment.intention.create, investment.intention.update, investment.group_decision.create, investment.group_decision.update, admin.user.read, admin.department.manage]
      parameters:
        - name: unitId
          in: query
          schema: { type: string, format: uuid }
        - name: departmentId
          in: query
          schema: { type: string, format: uuid }
        - name: permission
          in: query
          schema: { type: string, maxLength: 128 }
      responses: *okResponses
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ltc_session
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 64, maxLength: 64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    ResetPasswordInput:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password: { type: string, minLength: 12, maxLength: 128, format: password }
    EntityStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    UnitInput:
      type: object
      additionalProperties: false
      required: [code, name, status, sortOrder]
      properties:
        parentId: { type: [string, 'null'], format: uuid }
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 200 }
        status: { $ref: '#/components/schemas/EntityStatus' }
        sortOrder: { type: integer, minimum: 0 }
    UnitUpdateInput:
      allOf:
        - $ref: '#/components/schemas/UnitInput'
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 1 }
    DepartmentInput:
      type: object
      additionalProperties: false
      required: [unitId, code, name, status, sortOrder]
      properties:
        unitId: { type: string, format: uuid }
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 200 }
        departmentHeadUserId: { type: [string, 'null'], format: uuid }
        supervisingLeaderUserId: { type: [string, 'null'], format: uuid }
        status: { $ref: '#/components/schemas/EntityStatus' }
        sortOrder: { type: integer, minimum: 0 }
    DepartmentUpdateInput:
      allOf:
        - $ref: '#/components/schemas/DepartmentInput'
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 1 }
    PositionInput:
      type: object
      additionalProperties: false
      required: [departmentId, code, name, status, sortOrder]
      properties:
        departmentId: { type: string, format: uuid }
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 200 }
        status: { $ref: '#/components/schemas/EntityStatus' }
        sortOrder: { type: integer, minimum: 0 }
    PositionUpdateInput:
      allOf:
        - $ref: '#/components/schemas/PositionInput'
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 1 }
    RoleInput:
      type: object
      additionalProperties: false
      required: [code, name, status, permissionCodes]
      properties:
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 200 }
        description: { type: [string, 'null'], maxLength: 500 }
        status: { $ref: '#/components/schemas/EntityStatus' }
        permissionCodes:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 128 }
    RoleUpdateInput:
      allOf:
        - $ref: '#/components/schemas/RoleInput'
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 1 }
    UserInput:
      type: object
      additionalProperties: false
      required: [employeeNo, username, displayName, password, status, roleIds]
      properties:
        employeeNo: { type: string, maxLength: 64 }
        username: { type: string, maxLength: 100 }
        displayName: { type: string, maxLength: 100 }
        password: { type: string, minLength: 12, maxLength: 128, format: password }
        mobile: { type: [string, 'null'], maxLength: 32 }
        email: { type: [string, 'null'], maxLength: 200, format: email }
        departmentId: { type: [string, 'null'], format: uuid }
        positionId: { type: [string, 'null'], format: uuid }
        status: { $ref: '#/components/schemas/EntityStatus' }
        roleIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
    UserUpdateInput:
      allOf:
        - $ref: '#/components/schemas/UserInput'
        - type: object
          required: [version]
          properties:
            password: false
            version: { type: integer, minimum: 1 }
...